+++ frontend/src/app/dashboard/become-a-mechanic/_components/mechanic-verification.tsx
@@ -2,11 +2,12 @@ import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
 import { Badge } from "@/components/ui/badge"
 import { Clock, Lock, CheckCircle } from "lucide-react"
+import { ethers } from 'ethers';
 
 export default function MechanicVerification() {
 
 
-  
+
   return (
     <div className="container mx-auto p-4 space-y-6">
       <h1 className="text-3xl font-bold text-left mb-8 text-amber-700 border-amber-700 pb-4">Mechanic Verification</h1>
